Move interval object and tests into separate files
