+++ src/undate/__init__.py
@@ -1,6 +1,7 @@
 __version__ = "0.4.0.dev0"
 
 from undate.date import DatePrecision
-from undate.undate import Undate, UndateInterval
+from undate.undate import Undate, Calendar
+from undate.interval import UndateInterval
 
-__all__ = ["Undate", "UndateInterval", "DatePrecision", "__version__"]
+__all__ = ["Undate", "UndateInterval", "Calendar", "DatePrecision", "__version__"]
