+++ src/undate/converters/calendars/hebrew/transformer.py
@@ -1,6 +1,6 @@
 from lark import Transformer, Tree
 
-from undate.undate import Undate, Calendar
+from undate import Undate, Calendar
 
 
 class HebrewUndate(Undate):
