+++ src/undate/converters/calendars/hijri/converter.py
@@ -3,10 +3,10 @@ from typing import Union
 from convertdate import islamic  # type: ignore
 from lark.exceptions import UnexpectedCharacters
 
+from undate import Undate, UndateInterval
 from undate.converters.base import BaseCalendarConverter
 from undate.converters.calendars.hijri.parser import hijri_parser
 from undate.converters.calendars.hijri.transformer import HijriDateTransformer
-from undate.undate import Undate, UndateInterval
 
 
 class HijriDateConverter(BaseCalendarConverter):
